Fixed mobile on homepage/profile
