+++ client/src/components/MyWorkouts/myworkouts.css
@@ -63,6 +63,15 @@
     line-height: 1.6em;
   }
   
+  .mockQuery {
+    display: flex;
+    opacity: .2;
+  }
+
+  .mockQuery .mock-query-card {
+    margin: 15px;
+  }
+
   @media only screen and (max-width: 915px) {  
 /* Query cards  */
